add QT viewer
